Cut down on the generated rules.
Skip FRules that have things like (x + 0 = y) in the assumptions.
This is OK because when we add "x + 0 = y", it will be replaced by "x = y"
so we should never really see "x + 0 = y" in the assumptions.

This implementation is a bit ad-hoc, it'd be nice to work out the details
of a more general scheme for this optimization.